Reformatting and preparing for debug mode
Created Helpers.py file and combined get_combinations.py, loading_animation.py, Constants.py, and Checks.py.
