+++ main/Exporter.py
@@ -10,8 +10,8 @@ import json
 import smtplib
 import datetime
 import platform
-from .loading_animation import Loader
-from .Constants import bcolors, removeList, remove_file_by_extension
+
+from .Helpers import bcolors, Loader
 from .Metadata import createCardanoMetadata, createSolanaMetaData, createErc721MetaData
